--- app/src/main/java/ru/gb/veber/paadlesson1/model/DataModel.kt
@@ -1,15 +1,0 @@
-package ru.gb.veber.paadlesson1.model
-
-import com.google.gson.annotations.SerializedName
-
-class DataModel(
-    @field:SerializedName("text") val text: String?,
-    @field:SerializedName("meanings") val meanings: List<Meanings>?,
-)
-
-class Meanings(
-    @field:SerializedName("translation") val translation: Translation?,
-    @field:SerializedName("imageUrl") val imageUrl: String?,
-)
-
-class Translation(@field:SerializedName("text") val translation: String?)
