+++ app/src/main/java/ru/gb/veber/paadlesson1/model/data/AppState.kt
@@ -1,6 +1,5 @@
-package ru.gb.veber.paadlesson1
+package ru.gb.veber.paadlesson1.model.data
 
-import ru.gb.veber.paadlesson1.model.DataModel
 
 sealed class AppState {
     data class Success(val data: List<DataModel>?) : AppState()
