+++ app/src/main/java/ru/gb/veber/paadlesson1/model/data/DataModel.kt
@@ -1,0 +1,8 @@
+package ru.gb.veber.paadlesson1.model.data
+
+import com.google.gson.annotations.SerializedName
+
+class DataModel(
+    @field:SerializedName("text") val text: String?,
+    @field:SerializedName("meanings") val meanings: List<Meanings>?,
+)
