+++ app/src/main/java/ru/gb/veber/paadlesson1/model/data/Translation.kt
@@ -1,0 +1,5 @@
+package ru.gb.veber.paadlesson1.model.data
+
+import com.google.gson.annotations.SerializedName
+
+class Translation(@field:SerializedName("text") val translation: String?)
