+++ app/src/main/java/ru/gb/veber/paadlesson1/model/datasources/DataSource.kt
@@ -1,0 +1,7 @@
+package ru.gb.veber.paadlesson1.model.datasources
+
+import io.reactivex.Observable
+
+interface DataSource<T> {
+    fun getData(word: String): Observable<T>
+}
