+++ app/src/main/java/ru/gb/veber/paadlesson1/model/datasources/DataSourceLocal.kt
@@ -1,0 +1,13 @@
+package ru.gb.veber.paadlesson1.model.datasources
+
+import io.reactivex.Observable
+import ru.gb.veber.paadlesson1.model.data.DataModel
+
+class DataSourceLocal(
+    private val remoteProvider: RoomDataBaseImplementation =
+        RoomDataBaseImplementation(),
+) :
+    DataSource<List<DataModel>> {
+    override fun getData(word: String): Observable<List<DataModel>> =
+        remoteProvider.getData(word)
+}
