+++ app/src/main/java/ru/gb/veber/paadlesson1/model/datasources/RepositoryImplementation.kt
@@ -1,8 +1,8 @@
-package ru.gb.veber.paadlesson1.model
+package ru.gb.veber.paadlesson1.model.datasources
 
 import io.reactivex.Observable
-import ru.gb.veber.paadlesson1.myinterface.DataSource
-import ru.gb.veber.paadlesson1.myinterface.Repository
+import ru.gb.veber.paadlesson1.model.data.DataModel
+import ru.gb.veber.paadlesson1.model.repository.Repository
 
 class RepositoryImplementation(private val dataSources: DataSource<List<DataModel>>) :
     Repository<List<DataModel>> {
