+++ app/src/main/java/ru/gb/veber/paadlesson1/model/repository/Repository.kt
@@ -1,0 +1,7 @@
+package ru.gb.veber.paadlesson1.model.repository
+
+import io.reactivex.Observable
+
+interface Repository<T> {
+    fun getData(word: String): Observable<T>
+}
