+++ app/src/main/java/ru/gb/veber/paadlesson1/presenter/Interactor.kt
@@ -1,0 +1,7 @@
+package ru.gb.veber.paadlesson1.presenter
+
+import io.reactivex.Observable
+
+interface Interactor<T> {
+    fun getData(word: String, fromRemoteSources: Boolean): Observable<T>
+}
