+++ app/src/main/java/ru/gb/veber/paadlesson1/rx/SchedulerProvider.kt
@@ -1,11 +1,9 @@
-package ru.gb.veber.paadlesson1.presenter
+package ru.gb.veber.paadlesson1.rx
 
 import io.reactivex.Scheduler
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
-import ru.gb.veber.paadlesson1.presenter.ISchedulerProvider
 
-//In the sake of testing
 class SchedulerProvider : ISchedulerProvider {
 
     override fun ui(): Scheduler = AndroidSchedulers.mainThread()
