+++ app/src/main/java/ru/gb/veber/paadlesson1/view/base/BaseActivity.kt
@@ -1,10 +1,9 @@
-package ru.gb.veber.paadlesson1.view
+package ru.gb.veber.paadlesson1.view.base
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import ru.gb.veber.paadlesson1.AppState
-import ru.gb.veber.paadlesson1.myinterface.Presenter
-import ru.gb.veber.paadlesson1.myinterface.View
+import ru.gb.veber.paadlesson1.model.data.AppState
+import ru.gb.veber.paadlesson1.presenter.Presenter
 
 abstract class BaseActivity<T : AppState> : AppCompatActivity(), View {
