+++ app/src/main/java/ru/gb/veber/paadlesson1/view/base/View.kt
@@ -1,0 +1,7 @@
+package ru.gb.veber.paadlesson1.view.base
+
+import ru.gb.veber.paadlesson1.model.data.AppState
+
+interface View {
+    fun renderData(appState: AppState)
+}
