+++ app/src/main/java/ru/gb/veber/paadlesson1/view/main/MainAdapter.kt
@@ -1,4 +1,4 @@
-package ru.gb.veber.paadlesson1.view
+package ru.gb.veber.paadlesson1.view.main
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import ru.gb.veber.paadlesson1.R
-import ru.gb.veber.paadlesson1.model.DataModel
+import ru.gb.veber.paadlesson1.model.data.DataModel
 
 
 class MainAdapter(
