+++ app/src/test/java/ru/gb/veber/paadlesson1/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package ru.gb.veber.paadlesson1
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
